feat(CLI): Allows configuration of changelogSections via CLI and provides option for using a configuration file for options.
- `changelog-sections` will now be parsed similar to the feature found in `release-please-action`
- Since this option is likely to be a bit unwieldy in CLI contexts, the CLI now accepts a `--config` option to supply options from a JSON configuration file.

- A `--no-operation/--no-op` flag has been added to allow short-circuiting of Yargs `.command` handlers to appease tests (`nock`)

see: googleapis/release-please#511
