+++ test/fixtures/example-config.json
@@ -1,0 +1,5 @@
+{
+  "package-name": "cli-package--config",
+  "repo-url": "googleapis/release-please-cli",
+  "changelog-sections": "[{\"type\":\"feat\",\"section\":\"Features\"},{\"type\":\"fix\",\"section\":\"Bug Fixes\"},{\"type\":\"perf\",\"section\":\"Performance Improvements\"},{\"type\":\"deps\",\"section\":\"Dependencies\"},{\"type\":\"revert\",\"section\":\"Reverts\"},{\"type\":\"docs\",\"section\":\"Documentation\"},{\"type\":\"style\",\"section\":\"Styles\",\"hidden\":true},{\"type\":\"chore\",\"section\":\"Miscellaneous Chores\"},{\"type\":\"refactor\",\"section\":\"Code Refactoring\",\"hidden\":true},{\"type\":\"test\",\"section\":\"Tests\",\"hidden\":true},{\"type\":\"build\",\"section\":\"Build System\",\"hidden\":true},{\"type\":\"ci\",\"section\":\"Continuous Integration\",\"hidden\":true}]"
+}
